64bit fixes

+++ src/script/common/c_converter.cpp
@@ -545,10 +545,10 @@ std::string getstringfield_default(lua_State *L, int table,
 	return result;
 }
 
-int getintfield_default(lua_State *L, int table,
-		const char *fieldname, int default_)
+long getintfield_default(lua_State *L, int table,
+		const char *fieldname, long default_)
 {
-	int result = default_;
+	long result = default_;
 	getintfield(L, table, fieldname, result);
 	return result;
 }
